requirements.txt
streamlit
joblib
numpy
pandas
scikit-learn
pymongo
python-dotenv

